feat: get weather by clicking button City or Location, show weather broadcast on the screen
