+++ src/app/components/weather-forecast/weather-forecast.component.html
@@ -1,5 +1,58 @@
-<div *ngIf="data$ | async as data">
-  <app-day-forecast [today]="today" [data]="data"></app-day-forecast>
-  <app-weather-details [data]="data"></app-weather-details>
-</div>
+<mat-card
+  style="
+  display: flex;
+  flex-direction: column;">
 
+  <mat-card-header>
+    <mat-card-title style="margin: 1rem 0;">
+      <button mat-fab extended color="primary" (click)="getWeatherForCity()" style="margin-right: 1rem;">
+        <mat-icon>add_location</mat-icon>
+          Find city
+      </button>
+      <button mat-fab extended color="accent" (click)="getWeatherForLocation()">
+        <mat-icon>my_location</mat-icon>
+          Get weather for my location
+      </button>
+    </mat-card-title>
+    <mat-card-subtitle>
+      <mat-form-field *ngIf="search"
+        style="flex: 1 1 auto; margin: 1rem 0">
+        <input
+          type="text"
+          matInput
+          placeholder="Search places"
+          [formControl]="searchControl"
+          [matAutocomplete]="auto"
+        >
+
+        <mat-autocomplete
+          autoActiveFirstOption
+          #auto="matAutocomplete"
+          (optionSelected)='selectOption($event)'>
+            <mat-option *ngFor="let city of cityAutoSuggestions | async" [value]="city.name">
+              <mat-icon matTextPrefix>location_on</mat-icon>
+              <span>{{getCityDisplayName(city)}}</span>
+            </mat-option>
+        </mat-autocomplete>
+
+        <button
+          mat-icon-button
+          matSuffix
+          class="close-icon"
+          aria-label="Icon-button with close icon"
+          (click)="closeInput()">
+          <mat-icon>close</mat-icon>
+        </button>
+
+      </mat-form-field>
+    </mat-card-subtitle>
+  </mat-card-header>
+
+  <mat-card-content *ngIf="data$ | async as data">
+    <mat-card>
+      <app-day-forecast [today]="today" [data]="data"></app-day-forecast>
+      <app-weather-details [data]="data"></app-weather-details>
+    </mat-card>
+  </mat-card-content>
+
+</mat-card>
